IDE-3109 fix blade url and move test to maven.core.tests
